add data files tests (1/N)
